requirements.txt
pandas>=1.5
google-cloud-bigquery>=3.0.0
pyarrow>=11.0.0
numpy>=1.24
python-dateutil
